more code cleaning

+++ src/routes/(app)/dashboard/applications/(ops)/edit/+page.svelte
@@ -9,39 +9,39 @@
 	let application = data.application;
 
 	let isDeletable = false;
-	let deleting = false;
-	let editing = false;
+	let isExecutingAction = false;
 
-	const handleEditApplication = async () => {
+	const handleAction = async (action: 'DELETE' | 'EDIT') => {
 		if (!$user?.uid) return;
-		editing = true;
-		await applicationToFirestore($user.uid, application);
+		isExecutingAction = true;
+		if (action === 'DELETE') await deleteApplicationFromFirestore($user.uid, application.id);
+		if (action === 'EDIT') await applicationToFirestore($user.uid, application);
 		goto('/dashboard/applications/');
-		editing = false;
-	};
-
-	const handleDeleteApplication = async () => {
-		if (!$user?.uid) return;
-		deleting = true;
-		await deleteApplicationFromFirestore($user.uid, application.id);
-		goto('/dashboard/applications/');
-		deleting = false;
 	};
 </script>
 
 <h2>edit application</h2>
-<ApplicationForm {application} />
-{#if !editing}
-	<button type="button" on:click={handleEditApplication}>done</button>
-{/if}
-{#if !deleting}
+<ApplicationForm bind:application />
+{#if !isExecutingAction}
+	<button
+		type="button"
+		on:click={() => {
+			handleAction('EDIT');
+		}}>done</button
+	>
 	<form>
 		<label for="delete">i want to delete this application</label>
 		<input type="checkbox" name="delete-application" id="delete" bind:checked={isDeletable} />
-		<button type="button" on:click={handleDeleteApplication} disabled={!isDeletable}>delete</button>
+		<button
+			type="button"
+			on:click={() => {
+				handleAction('DELETE');
+			}}
+			disabled={!isDeletable}>delete</button
+		>
 	</form>
 {/if}
-{#if deleting || editing}
+{#if isExecutingAction}
 	<iconify-icon icon="line-md:loading-twotone-loop" />
 {/if}
